Suppress warnings during compilation

+++ src/Problem.cpp
@@ -49,7 +49,7 @@ double Problem::objective(Solution &x) {
     double cost = 0;
     int old = x_data.at(0);
 
-    for (int i=1; i < x_data.size(); i++) {
+    for (long unsigned int i=1;i<x_data.size();i++) {
         cost += m_data.at(old).at(x_data.at(i));
         old = x_data.at(i);
     }
@@ -64,8 +64,8 @@ double Problem::objective(Solution &x) {
 // Solution Problem::solve(...);
 
 void Problem::print() const {
-    for (int i=0; i < m_data.size(); i++) {
-        for (int j=0; j < m_data.at(i).size(); j++) {
+    for (long unsigned int i=0;i<m_data.size();i++) {
+        for (long unsigned int j=0;j<m_data.at(i).size();j++) {
             cout << m_data.at(i).at(j) << ' ';
         }
         cout << endl;
